requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
pymupdf
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
openai
langchain-groq
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
fastapi
uvicorn
rank_bm25
python-pptx
langchain-experimental